// src/apis/models/FireBytes_Unified_Wave_Pacs_PacsCards_Dtos_PacsCardEditDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FireBytes_Unified_Wave_Pacs_PacsCards_Dtos_PacsCardEditDto = {
    id?: string | null;
    cardFormatId?: string;
    cardNumber?: string | null;
    facilityCode?: number;
    startDateTime?: string | null;
    endDateTime?: string | null;
    isVirtualCard?: boolean;
    badgeTemplateId?: string | null;
    cardStatus?: string | null;
    assignedToPacsUserId?: string | null;
    pinNumber?: string | null;
    descriptions?: string | null;
};